% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate-extraction.R
\name{adjust_elipe}
\alias{adjust_elipe}
\title{Adjust Leap Day (Feb 29) to Feb 28}
\usage{
adjust_elipe(date)
}
\arguments{
\item{date}{A \code{Date} vector.}
}
\value{
A \code{Date} vector where all leap days (Feb 29) are moved to Feb 28.
}
\description{
This function checks for any dates that fall on February 29 and adjusts them back to February 28,
to avoid issues when performing date calculations (e.g., in non-leap years).
}
